Fix strange edge case + slight i18n tweak
When there would be precisely *one* picture game in the DB, it wouldn't
show up for anyone when they'd visit [[Special:PictureGameHome]], which
definitely wasn't the intended behavior.

Added new 'picturegame-empty-no-more' message, which is shown when the
user has played all available picture games; there is an obvious
distinction between "you've played all existing pic games" and "there are
no pic games at all". Nevertheless in both situation we want to encourage
the user to contribute to the pool of available picture games.

Change-Id: Id5fffecc8619dbb5ac27578615bc6ca3dcb268b0