Unbreak picture game creation
When did this break and why? I've no idea, your guess is as good as mine.
All I know for sure is that prior to this patch, on MW 1.34, while uploading images from Special:PictureGameHome was possible,
pressing "Start and Play!" would cause the data to *not* be inserted into the DB tables, effectively leaving the images orphan
and resulting in minor data loss (the picture game title would end up being lost, as would be the image association with said title).

Change-Id: I41ab4857b77757b8fd44515d09421fe057bc3b0f