Improve i18n-ability for the thresholds feature
Literally just copypasted from SocialProfile's UserProfile/includes/specials/SpecialUpdateProfile.php
with s/user-profile/picturegame/g.

Previously if there was one (or more) thresholds for picture game creation defined with the $wgCreatePictureGameThresholds global,
such as $wgCreatePictureGameThresholds = [ 'edits' => 5 ]; to require users to have five edits before being allowed to create
a picture game, "5 edits" would be shown in English in the picturegame-create-threshold-reason message regardless of the user's
UI language and this string was not translatable. This patch corrects that and makes the threshold reason strings translatable,
just like in SocialProfile.

Change-Id: I6832c8630eb62215818ee99fbe637049388b235e