Allow using an existing (already uploaded) image file when creating&editing picture games
It's finally possible to use images already uploaded to the wiki instead of being forced to upload a new image when creating a new picture game or editing an existing one via Special:PictureGameHome.

As per the comments, this is quite literally ext.CollaborationKit.hubtheme.js from CollaborationKit with minimal adjustments.
The biggest "change" is that I removed all the CK-specific color picker stuff, but obviously i18n message names and certain magic numbers were also changed.

Unlike with FanBoxes, PollNY and QuizGame, some more work had to be done here to account for the fact that PictureGame has *two* uploading forms, whereas the other aforementioned three only have one.
Additional headache was caused by the fact that when *creating* a new picture game, the two uploading forms are shown side-by-side...but this is not the case when editing an existing game, as then only one uploading form will be shown at a time, and the code needs to know which image is being edited, the first or the second one.

Change-Id: If3d1102a45c9882698d1b76c02697df094a22ddf